+++ appd-services/src/client/login.ts
@@ -3,24 +3,27 @@ import { HttpClient } from './http';
 import { LoginProvider } from './login-providers/spi';
 import { LDAPLoginProvider } from './login-providers/ldap';
 import { SAMLLoginProvider } from './login-providers/saml';
+import { LocalLoginProvider } from './login-providers/local';
 
 const providers = {
     'saml': SAMLLoginProvider,
-    'ldap': LDAPLoginProvider
+    'ldap': LDAPLoginProvider,
+    'local' : LocalLoginProvider
 } as any;
 
 
-function login(url:string, account:string, username:string, password:string):Promise<Session> {
+function login(url:string, account:string, username:string, password:string, method?:string):Promise<Session> {
     const http = new HttpClient({baseURL:url});
-    return http
-    .get(`/public-info?action=query-security-provider&account-name=${account}`)
-    .then(rsp => rsp.data.trim())
+    return (method 
+        ? Promise.resolve(method) 
+        : http.get(`/public-info?action=query-security-provider&account-name=${account}`)
+          .then(rsp => rsp.data.trim())
+    )
     .then(method => {
         try {
             const Provider = providers[method.toLowerCase()];
             return new Provider() as LoginProvider;
         } catch (e) {
-            console.error(e);
             throw new Error(`no provider found for method ${method}`)
         }
     })